Add temperature information to intro text
This commit adds a new feature which displays the current weather for Kent within
the intro text. As the weather conditions change, so do the icons that are displayed
next to the temperature.

The preferred unit of temperature is detected based on the user's locale, and
can be toggled between Celsius and Fahrenheit.